+++ CHANGELOG.md
@@ -1,10 +1,13 @@
 v2.17.0
 =======
+* Introduced converter for `char` primitive and `Character` class
 * Introduced new specifications:
   * `isEmpty`, `isNotEmpty` - these specifications filter out elements that have empty (not empty) collection of elements, that is defined under `path` in `@Spec` annotation.
+  * `Empty` - this specification filters for collections using `is empty` or `is not empty`, depending on the value of the parameter passed in (e.g. ` where customer.orders is empty`).
 
 v2.16.0
 =======
+* Introduced `EqualDay` specification which allows finding all records within particular date (day), ignoring time.
 * Added ability to set custom `Locale` during resolver registration:
   ```java
   @Override